improving liquid engine, starting to add some interpolation

+++ src/include/_c_mandelbrot_benchmark.c
@@ -2,10 +2,10 @@
 #define MAX_ITERATIONS 1000
 #define DIVERGENCE 10
 
-int _c_mandelbrot(double row, double col) {
-  double zrow = 0;
-  double zcol = 0;
-  double zrow_new, zcol_new;
+int _c_mandelbrot(float row, float col) {
+  float zrow = 0;
+  float zcol = 0;
+  float zrow_new, zcol_new;
   int iterations = 0;
   while (zrow * zrow + zcol * zcol <= DIVERGENCE &&
          iterations < MAX_ITERATIONS) {
